Move `as_ptr` and `as_mut_ptr` into a `utils` module.
This tidies up the top-level lib.rs.

While here, replace the manual implementations of `offsetof_sun_path`
with calls to the actual `offsetof_sun_path`, so that we have fewer
uses of `as_ptr`.
